refactor : error -> exception으로 디렉토리명 변경 refactor : commandType.java 에서 String input parmeter 명을 command로 변경 refactor : VoucherController, MemberController throws 명시 refactor : FixedVoucher, PercentVoucher 생성자 throws 명시
