+++ part1/src/main/java/com/programmers/part1/exception/voucher/PercentErrorException.java
@@ -1,4 +1,4 @@
-package com.programmers.part1.error.voucher;
+package com.programmers.part1.exception.voucher;
 
 // PercentVoucher에서 amount가 0이하 100초과 일 경우 처리
 public class PercentErrorException extends RuntimeException {
